# Conflicts:
#	web/src/main/java/org/geogebra/web/full/gui/dialog/DialogManagerW.java

+++ retex/renderer-base/src/main/java/com/himamis/retex/renderer/share/FramedBox.java
@@ -107,7 +107,7 @@ public class FramedBox extends Box {
 			g2.setStroke(graphics.createBasicStroke(thickness,
 					BasicStroke.CAP_BUTT, BasicStroke.JOIN_MITER));
 		} else {
-			float[] dashes = new float[] { (float) dashdash,
+			double[] dashes = new double[] { (float) dashdash,
 					(float) (dashlength - dashdash) };
 			g2.setStroke(graphics.createBasicStroke(thickness, dashes));
 		}
